use floats

+++ include/vertex.hpp
@@ -20,7 +20,7 @@ namespace vertex
     {
         std::array<float, 3> position;
         std::array<float, 4> color;
-        std::array<uint16_t, 2> texcoord;
+        std::array<float, 2> texcoord;
     };
 
     static inline GPU_Primitive_t GetMode(TriangleIndexMode mode)
